############################################################### Move NFM Rules to target Layer ###############################################################
Test scenario:
1- Check that NFM rules are enabled when MAP-T is up and disabled when it's down

Test results: Ok
